backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
ezdxf
requests
python-dotenv
pydantic
apscheduler
